user_service/requirements.txt
pyjwt==1.7.1
flask==1.1.1